[WMVCORE] Sync with Wine Staging 4.18. CORE-16441
